APIのディレクトリ変更、変数名の修正
